Use typst flake from upstream
In order to get the latest, one should run `nix flake update`. This is
the standard workflow for a Nix flake, as one needs to be explicit about
what software one uses. Because the typst github is listed as a flake
input, the explicit reference to the version is kept in flake.lock. This
is preferable to the user having to explicitly update the package hash
in flake.nix.